pyproject.toml
[project]
name = "lumina-agent-templates"
version = "0.1.0"
description = "Template sync tooling for Lumina agent templates"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.21.1",
  "pyyaml>=6.0.3",
]

[project.scripts]
lumina-templates = "template_sync.cli:main"

[project.optional-dependencies]
test = [
  "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["template_sync"]
